+++ changelog.md
@@ -4,6 +4,10 @@
 
 - pdf : minification des fichiers .mjs, suppression de la librairie wasm, suppression répertoire mobile (inutilisé)
 
+#### Modification interne
+
+- up.php :  cleanup checkfiles
+
 ## 06/11/2025 - version 5.4.1
 
 #### Modifications actions